Update patch set 1
Patch Set 1:

It doesn't do anything, really.

https://github.com/CyanogenMod/android_kernel_cyanogen_msm8916/commit/345ef1c38f540df63113c9cbcaebc198f505f0a4

Long story short, when max_freq_hysteresis is the same value as min_sample_time, they literally do the same thing.

The setting in cyanogen msm8916-common is redundant too, and can be removed. It was actually removed in msm8939-common, but not on msm8916-common yet.

Patch-set: 1