msm8916-common: Correct fingerprint for ZE551KL
The build fingerprint is being constructed in a way that
will be valid for one of the unified device variants, but
will almost certainly be incorrect for all of the others.

Modify it so that the build fingerprint constructed for
the ZE551KL hardware variant will actually correspond to
ASUS release V21.40.1220.2181, while leaving all other
hardware variants unmodified.

I cannot vouch for what the correct signatures ought to be
in the other variants, so they will be left unmodified.

Change-Id: I0f5e7ba03359f3c1f79cd576d28bd3d8ef1eef33